requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
defusedxml==0.7.1
Django==4.1.6
django-allauth==0.52.0
django-ckeditor==6.5.1
django-filter==22.1
django-js-asset==2.0.0
idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.1
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
urllib3==1.26.14
